+++ FrontendRSReact/src/assets/navigation.css
@@ -61,6 +61,7 @@ nav .active {
 .logo-container {
   display: flex;
   align-items: center;
+  background-color: #CDE0C9;
 }
 
 .logo {
