+++ FrontendRSReact/src/assets/productpages.css
@@ -5,8 +5,7 @@
   right: 0;
   z-index: 999;
   background-color: #fff;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-  height: calc(100vh - 50px); /* Adjust the height based on your needs */
+  /* box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1); */
 
 }
