+++ FrontendRSReact/src/assets/tabsnav.css
@@ -6,7 +6,7 @@
     position: sticky;
     display: flex;
     justify-content: center;
-    gap: 200px;
+    gap: 50px;
     box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2); /* Perbarui efek box-shadow */
   }
