+++ FrontendRSReact/src/components/Footer.css
@@ -1,15 +1,51 @@
 .footer {
-  position: fixed;
+  
+    background-color: #f8f9fa;
+    padding: 10px;
+    text-align: center;
+    position: relative;
+    margin-top: 50px; 
+
+  /* background-color: #ffffff;
+  font-size: smaller;
   bottom: 0;
   left: 0;
-  background-color: #ffffff;
-  color: #a8a8a8;
-  font-size: 10px;
-  font-weight: bold;
   width: 100%;
-  height: 20px;
-  display: flex;
+  height: 90px;
+  z-index: 999;
+  position: relative;
+  display:flex;
+  justify-content: space-between;
   align-items: center;
-  justify-content: flex-start;
-  padding-left: 10px;
+  padding: 0 10px;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2); */
+}
+
+.footer-text-bold {
+  font-weight: bold;
+  font-size:15px;
+  color: #000000;
+  /* margin-right: 200px; */
+}
+
+.footer-text-small {
+  font-size:10px;
+  color: #000000;
+  /* margin-right: 300px; */
+}
+
+.footer-social-icons a {
+  color: #000;
+  font-size: 20px;
+  margin-right: 10px;
+}
+
+.footer-text {
+  color: #000000;
+  font-size: 10px;
+}
+
+.footer-buttons .btn {
+  margin-right: 1px;
+  font-size: smaller;
 }
